Module sos21_api_server::handler::model

API model types.

API's model types are implemented independently from sos21_use_case::model. We never use data transfer object from sos21-use-case directly as API's model types, because it's horribly not good that changes in sos21-use-case DTO affects the public interface, or specification, of the API server. In the other words, we want to keep the public interface of the API server self-contained in sos21-api-server.

Modules

date_time

distributed_file

file

file_distribution

file_sharing

form

form_answer

pending_project

project

project_query

registration_form

registration_form_answer

serde

user

user_invitation